requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
asyncer==0.0.17
attrs==25.4.0
audioop-lts==0.2.2
bidict==0.23.1
certifi==2026.2.25
cffi==2.0.0
chainlit==2.10.0
charset-normalizer==3.4.6
chevron==0.14.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
cuid==0.4
dataclasses-json==0.6.7
deprecated==1.3.1
distro==1.9.0
docstring-parser==0.17.0
dotenv==0.9.9
fastapi==0.135.1
filelock==3.25.2
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.2.0
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==1.7.1
idna==3.11
importlib-metadata==8.7.1
inflection==0.5.1
jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.12
langchain-anthropic==1.3.5
langchain-core==1.2.19
langgraph==1.1.2
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.11
langsmith==0.7.17
lazify==0.4.0
literalai==0.1.201
markdown-it-py==4.0.0
markupsafe==3.0.3
marshmallow==3.26.2
mcp==1.26.0
mdurl==0.1.2
multidict==6.7.1
mypy-extensions==1.1.0
nest-asyncio==1.6.0
oauthlib==3.3.1
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-instrumentation==0.61b0
opentelemetry-instrumentation-agno==0.53.0
opentelemetry-instrumentation-alephalpha==0.53.0
opentelemetry-instrumentation-anthropic==0.53.0
opentelemetry-instrumentation-bedrock==0.53.0
opentelemetry-instrumentation-chromadb==0.53.0
opentelemetry-instrumentation-cohere==0.53.0
opentelemetry-instrumentation-crewai==0.53.0
opentelemetry-instrumentation-google-generativeai==0.53.0
opentelemetry-instrumentation-groq==0.53.0
opentelemetry-instrumentation-haystack==0.53.0
opentelemetry-instrumentation-lancedb==0.53.0
opentelemetry-instrumentation-langchain==0.53.0
opentelemetry-instrumentation-llamaindex==0.53.0
opentelemetry-instrumentation-logging==0.61b0
opentelemetry-instrumentation-marqo==0.53.0
opentelemetry-instrumentation-mcp==0.53.0
opentelemetry-instrumentation-milvus==0.53.0
opentelemetry-instrumentation-mistralai==0.53.0
opentelemetry-instrumentation-ollama==0.53.0
opentelemetry-instrumentation-openai==0.53.0
opentelemetry-instrumentation-openai-agents==0.53.0
opentelemetry-instrumentation-pinecone==0.53.0
opentelemetry-instrumentation-qdrant==0.53.0
opentelemetry-instrumentation-redis==0.61b0
opentelemetry-instrumentation-replicate==0.53.0
opentelemetry-instrumentation-requests==0.61b0
opentelemetry-instrumentation-sagemaker==0.53.0
opentelemetry-instrumentation-sqlalchemy==0.61b0
opentelemetry-instrumentation-threading==0.61b0
opentelemetry-instrumentation-together==0.53.0
opentelemetry-instrumentation-transformers==0.53.0
opentelemetry-instrumentation-urllib3==0.61b0
opentelemetry-instrumentation-vertexai==0.53.0
opentelemetry-instrumentation-voyageai==0.53.0
opentelemetry-instrumentation-watsonx==0.53.0
opentelemetry-instrumentation-weaviate==0.53.0
opentelemetry-instrumentation-writer==0.53.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opentelemetry-semantic-conventions-ai==0.4.15
opentelemetry-util-http==0.61b0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.13.1
pygments==2.19.2
pyjwt==2.12.1
pyparsing==3.3.2
python-dotenv==1.2.2
python-engineio==4.13.1
python-multipart==0.0.22
python-socketio==5.16.1
pywin32==311
pyyaml==6.0.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
shellingham==1.5.4
simple-websocket==1.1.0
sniffio==1.3.1
sse-starlette==3.3.2
starlette==0.52.1
syncer==2.0.3
tenacity==9.1.4
tokenizers==0.22.2
tomli==2.4.0
tqdm==4.67.3
traceloop-sdk==0.53.0
typer==0.24.1
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
uritemplate==4.2.0
urllib3==2.6.3
uuid-utils==0.14.1
uvicorn==0.41.0
watchfiles==1.1.1
wrapt==1.17.3
wsproto==1.3.2
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
